Refactor pinot-batch-ingestion job specs to pinot-spi
