--- pinot-plugins/pinot-batch-ingestion/pinot-batch-ingestion-base/src/main/java/org/apache/pinot/ingestion/common/PinotClusterSpec.java
@@ -1,45 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.pinot.ingestion.common;
-
-/**
- * PinotClusterSpec defines the Pinot Cluster Access Point.
- */
-public class PinotClusterSpec {
-
-  /**
-   * Controller URI is used to fetch table/schema information and data push.
-   * E.g. http://localhost:9000
-   */
-  private String _controllerURI;
-
-  public String getControllerURI() {
-    return _controllerURI;
-  }
-
-  /**
-   * Controller URI is used to fetch table/schema information and data push.
-   * E.g. http://localhost:9000
-   *
-   * @param controllerURI
-   */
-  public void setControllerURI(String controllerURI) {
-    _controllerURI = controllerURI;
-  }
-}
