--- pinot-plugins/pinot-batch-ingestion/pinot-batch-ingestion-base/src/main/resources/segmentTarPushJobSpec.yaml
@@ -1,40 +1,0 @@
-#
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-#
-
-jobType: SegmentTarPush
-inputDirURI: 'file:///path/to/input'
-includeFileNamePattern: 'glob:**/*.parquet'
-excludeFileNamePattern: 'glob:**/*.avro'
-outputDirURI: 'file:///path/to/output'
-overwriteOutput: true
-pinotFSSpecs:
-  - scheme: file
-    className: org.apache.pinot.spi.filesystem.LocalPinotFS
-recordReaderSpec:
-  dataFormat: 'parquet'
-  className: 'org.apache.pinot.parquet.data.readers.ParquetRecordReader'
-tableSpec:
-  tableName: 'myTable'
-  schemaURI: 'http://localhost:9000/tables/myTable/schema'
-  tableConfigURI: 'http://localhost:9000/tables/myTable'
-pinotClusterSpecs:
-  - controllerURI: 'localhost:9000'
-pushJobSpec:
-  pushAttempts: 2
-  pushRetryIntervalMillis: 1000
